src/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==35.0.0
Django==4.2.11
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
install==1.3.5
phonenumbers==8.12.46
Pillow==9.4.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.10.0
tzdata==2024.1